Created Category Component,
